# src/main/java/com/loudlyapp/user/UserController.java
package com.loudlyapp.user;

import lombok.AllArgsConstructor;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Optional;

@RestController
@RequestMapping("/api/users")
@AllArgsConstructor
public class UserController {

    private final UserService userService; 

    @GetMapping
    public List<UserDTO> getUsers() {
        return userService.getAllUsers();
    }

    @GetMapping("/{userId}")
    public Optional<UserDTO> getUser(@PathVariable Long userId) {
        Optional<UserDTO> userDTO = userService.findById(userId);
        return userDTO;
    }

    @DeleteMapping()
    public void deleteAllUsers() {
        userService.deleteAll();
    }

    @PostMapping
    public UserDTO createUser(@RequestBody UserDTO userDTO) {
        return userService.save(userDTO);
    }


    @DeleteMapping("/{userId}")
    public void deleteUserById(@PathVariable Long userId) {
        userService.deleteUserById(userId);
    }


    @PutMapping("/{userId}")
    public UserDTO updateUser(@PathVariable Long userId, @RequestBody UserDTO userDTO) {
        UserDTO updatedUserDTO = userService.updateUserDTO(userId, userDTO);
        return updatedUserDTO;
    }
}